room_api

+++ src/main/java/com/qttd/repositories/CategoryRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.qttd.entities.CategoryEntity;
 
 @Repository
-public interface CategoryRepository extends CrudRepository<CategoryEntity, Integer> {
+public interface CategoryRepository extends CrudRepository<CategoryEntity, Long> {
 	List<CategoryEntity> findByCategoryName(String name);
 }
